submission to dr brain love (PROTOTYPE)
The inspiration for this model is the brain's large scale networks as
observed with fMRI and MEG. The code defines paths through the brain
model that mimic how activation flows during vision. The current code
displays (roughly) the visual ventral stream.

However: it is not a complete product. This is a basic prototype that I
need to extend in a bunch of ways. The to do list includes:

-adding other streams (obviously)
-adding spatial spread with some randomness;
-adding color gradient in multiple directions to make the patterns look
more like activation, which has a gaussian spread around the center
-cleaning up the very clunky code

I apologize that I just did not have enough time to finish it.

So I'm submitting this right now but I really hope you would consider
displaying an amended version if I could have another week or so to work
on it...